+++ src/pages/foreign/index.tsx
@@ -3,7 +3,7 @@
 import SidebarLayout from '@/components/layout/sidebarLayout';
 import Spacing from '@/components/container/spacing/spacing';
 import CategoryCarousel from '@/components/carousel/categoryCarousel';
-import MainCategoryBookList from '@/components/list/categoryBookList/mainCategoryBookList';
+import MainCategoryBookList from '@/components/container/categoryBookList/mainCategoryBookList';
 import EventSection from '@/components/container/eventSection/eventSection';
 
 import { responsive } from '@/utils/checkResponsiveEnv';
